# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi BKPSDM Katingan",
        "description": "Ini adalah dokumentasi API yang digunakan untuk website BKPSDM Katingan.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/galeris": {
            "get": {
                "tags": [
                    "Galeri"
                ],
                "summary": "Menampilkan daftar semua gambar di galeri",
                "operationId": "getGaleriList",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Galeri"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Galeri"
                ],
                "summary": "Menambah gambar baru ke galeri (memerlukan token)",
                "operationId": "storeGaleri",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gambar berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Galeri"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/galeris/{id}": {
            "get": {
                "tags": [
                    "Galeri"
                ],
                "summary": "Menampilkan detail satu gambar galeri",
                "operationId": "getGaleriById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Galeri"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Galeri"
                ],
                "summary": "Menghapus gambar dari galeri (memerlukan token)",
                "operationId": "deleteGaleri",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Gambar berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Galeri": {
                "title": "Galeri",
                "description": "Model untuk data galeri",
                "properties": {
                    "id": {
                        "description": "ID unik dari item galeri",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Judul atau caption gambar",
                        "type": "string",
                        "example": "Kegiatan Pelatihan Dasar"
                    },
                    "image": {
                        "description": "Path relatif ke file gambar",
                        "type": "string",
                        "example": "galeri/image.jpg"
                    },
                    "created_at": {
                        "description": "Waktu upload gambar",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint /api/login. Contoh: 1|aBcDeFgHiJkLmNoPqRsTuVwXyZ",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}